Pipeline Page Crud and Toast Error Messages

+++ components/modals/DeleteConfirmationModal.tsx
@@ -2,14 +2,17 @@
 
 import styles from './delete-confirmation.module.css';
 import { useTransition, useState, useEffect } from 'react';
-import { deleteEnvironment } from "@/lib/actions/environments";
+import { type FormState } from '@/lib/types';
 
-interface DeleteConfirmationProps {
+interface DeleteConfirmationModalProps {
   id: string;
+  recordLabel: string;
   onDelete: () => void;
   onDeleteClose: () => void;
+  deleteRecord: (id: string) => Promise<FormState>;
+  onError: (message: string) => void;
 }
-export default function DeleteConfirmation({ id, onDelete, onDeleteClose }: DeleteConfirmationProps) {
+export default function DeleteConfirmationModal({ id, recordLabel, onDelete, onDeleteClose, onError, deleteRecord }: DeleteConfirmationModalProps) {
   const [_isDeleteTransitionPending, startDeleteTransition] = useTransition();
   const [ready, setReady] = useState(false);
 
@@ -22,14 +25,19 @@ export default function DeleteConfirmation({ id, onDelete, onDeleteClose }: Dele
   }, [])
 
   const handleDelete = () => startDeleteTransition(async () => {
-    await deleteEnvironment(id);
-    onDelete();
+    const deletedRecord = await deleteRecord(id);
+    if (deletedRecord.status === 'success') {
+      onDelete();
+    }
+    else if (deletedRecord.status === 'error') {
+      onError(deletedRecord.message);
+    }
   });
 
   return (
     <div className={styles["delete-overlay"]} style={{ backdropFilter: 'blur(5px)', WebkitBackdropFilter: 'blur(5px)' }} onClick={onDeleteClose}>
       <div className={styles["delete-container"]} onClick={e => e.stopPropagation()}>
-        <p>Delete this Environment?</p>
+        <p>Delete this {recordLabel}?</p>
         <div className={styles['btn-group']}>
           <button className={styles["cancel-btn"]} type="button" onClick={onDeleteClose}>Cancel</button>
           {!ready ?
